chore(PocketIC): make external types tests manual
