+++ rs/pocket_ic_server/src/lib.rs
@@ -34,8 +34,9 @@
 //! The start state is a dedicated state that always exists independent of which computations have
 //! been carried out. A state which has no outcoming computations is called a leaf.
 
-mod external_canister_types;
 mod nonmainnet_features;
+
+pub mod external_canister_types;
 pub mod pocket_ic;
 pub mod state_api;
